cleaning
cleaned up stat displays(rounded display values into ints)
added a description GUI
implemented object and enemy spawn on field, essentially a prototype for level  design

+++ Assets/gameScript/Combat/TB/tbEnemyAI.cs
@@ -45,7 +45,7 @@ public class tbEnemyAI : Turn
         Debug.Log("enemy has moved");
         PlayerStat.TakeDamage(es.attack);
        // Debug.Log("enemy dealt " + es.attack + "damage");
-        MoveManager.root.Q<UnityEngine.UIElements.TextElement>("log-text").text = ("Enemy" + " dealt " + es.attack + "damage");
+        MoveManager.root.Q<UnityEngine.UIElements.TextElement>("log-text").text = ("Enemy" + " dealt " + (int)es.attack + "damage");
         MoveManager.Deq();
     }
 }
